Update the build process to determine versioning using git tags
The autogen.sh script uses git tags to determine the version
of the project. If the distribution tarball is used and no git
repository is available then the VERSION file is used.

This also changes the versioning schema to simple X.Y for the
master branch releases and X.Y.0.C.GITID for master nightlies.

Using dash/minus in the upstream version was bad, because when
the downstream RPM is created, it's version should be based on
the full upstream version of used tarball. But RPM does not allow
minus sign in the Version field and the Release field is supposed
to track updates to spec file and sidecar patches only..

Change-Id: Iacfec13ee13abf05c781d820e865b2d4d54398ac
Signed-off-by: Martin Sivak <msivak@redhat.com>